fix api issues, make order number as url for all events, add specs
