+++ app/services/spree/intercom/events/order/place_service.rb
@@ -33,19 +33,4 @@ class Spree::Intercom::Events::Order::PlaceService < Spree::Intercom::BaseServic
       }
     }
   end
-
-  private
-
-    def order_url
-      Spree::Core::Engine.routes.url_helpers.order_url(@order, host: host_name, protocol: protocol_name)
-    end
-
-    def host_name
-      Rails.application.routes.default_url_options[:host].presence || "localhost:3000"
-    end
-
-    def protocol_name
-      Rails.application.routes.default_url_options[:protocol].presence || "http"
-    end
-
 end
