+++ spec/services/spree/intercom/events/coupon/apply_service_spec.rb
@@ -5,6 +5,7 @@ RSpec.describe Spree::Intercom::Events::Coupon::ApplyService, type: :service do
   let!(:user) { create(:user) }
   let!(:order) { create(:order, user_id: user.id) }
   let!(:coupon) { Spree::PromotionHandler::Coupon.new(order) }
+  let!(:order_url) { Spree::Core::Engine.routes.url_helpers.order_url(order, host: 'localhost:3000', protocol: 'http') }
 
   let!(:options) {
     {
@@ -21,7 +22,10 @@ RSpec.describe Spree::Intercom::Events::Coupon::ApplyService, type: :service do
       created_at: options[:time],
       user_id: user.intercom_user_id,
       metadata: {
-        order_number: order.number,
+        order_number: {
+          url: order_url,
+          value: order.number
+        },
         code: options[:code]
       }
     }
