+++ spec/services/spree/intercom/events/line_item/delete_service_spec.rb
@@ -6,6 +6,7 @@ RSpec.describe Spree::Intercom::Events::LineItem::DeleteService, type: :service 
   let!(:order) { line_item.order }
   let!(:user) { line_item.order.user }
   let!(:variant) { line_item.variant }
+  let!(:order_url) { Spree::Core::Engine.routes.url_helpers.order_url(order, host: 'localhost:3000', protocol: 'http') }
 
   let!(:options) {
     {
@@ -23,7 +24,10 @@ RSpec.describe Spree::Intercom::Events::LineItem::DeleteService, type: :service 
       created_at: options[:time],
       user_id: user.intercom_user_id,
       metadata: {
-        order_number: order.number,
+        order_number: {
+          url: order_url,
+          value: order.number
+        },
         product: options[:name],
         sku: options[:sku]
       }
@@ -46,8 +50,8 @@ RSpec.describe Spree::Intercom::Events::LineItem::DeleteService, type: :service 
       expect(event_service.instance_variable_get(:@time)).to eq(options[:time])
     end
 
-    it 'is expected to set @order_number' do
-      expect(event_service.instance_variable_get(:@order_number)).to eq(options[:order_number])
+    it 'is expected to set @order' do
+      expect(event_service.instance_variable_get(:@order)).to eq(order)
     end
 
     it 'is expected to set @sku' do
@@ -91,4 +95,5 @@ RSpec.describe Spree::Intercom::Events::LineItem::DeleteService, type: :service 
     end
   end
 
+  it_behaves_like 'order_url'
 end
