+++ spec/services/spree/intercom/events/shipment/ship_service_spec.rb
@@ -5,6 +5,7 @@ RSpec.describe Spree::Intercom::Events::Shipment::ShipService, type: :service do
   let!(:user) { create(:user) }
   let!(:order) { create(:order, user_id: user.id) }
   let!(:shipment) { create(:shipment, state: 'ready', order: order) }
+  let!(:order_url) { Spree::Core::Engine.routes.url_helpers.order_url(order, host: 'localhost:3000', protocol: 'http') }
 
   let!(:options) {
     {
@@ -20,7 +21,10 @@ RSpec.describe Spree::Intercom::Events::Shipment::ShipService, type: :service do
       created_at: shipment.updated_at,
       user_id: user.intercom_user_id,
       metadata: {
-        order_number: order.number,
+        order_number: {
+          url: order_url,
+          value: order.number
+        },
         shipment_number: shipment.number
       }
     }
@@ -79,4 +83,5 @@ RSpec.describe Spree::Intercom::Events::Shipment::ShipService, type: :service do
     end
   end
 
+  it_behaves_like 'order_url'
 end
